型定義をtypesファイルに切り出す
